% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{appg_benefits}
\alias{appg_benefits}
\title{Fetch details on APPG benefits in kind}
\usage{
appg_benefits(appg = NA, category = NA, source = NA,
  low_value_from = NA, low_value_to = NA, high_value_from = NA,
  high_value_to = NA, received_from = NA, received_to = NA,
  registered_from = NA, registered_to = NA)
}
\arguments{
\item{appg}{A character string of an APPG title written within a pair of single
qoute or double qoutes. Can be a vector with more than one element. By default
all APPG titles are selected.}

\item{category}{A characer string of a APPG type written within a pair of single
qoute or double qoutes. APPG types available: Country, Subject, Club. Can be a
vector with more than one element. By default all APPG types are selected.}

\item{source}{A character string of the source of benefit in kind written
within a pair of single qoute or double qoutes. Can be a vector with more than
one element. By default all sources are selected.}

\item{low_value_from}{A non-negative numeric indicating the lower value range
of the benefit in kind. By default the minimum value in the data is used.}

\item{low_value_to}{A non-negative numeric indicating the lower value range
of the benefit in kind. By default the maximum value if the data is used.}

\item{high_value_from}{A non-negative numeric indicating the higher value range
of the benefit in kind. By default the minimum value in the data is used.}

\item{high_value_to}{A non-negative numeric indicating the higher value range
of the benefit in kind. By default the maximum value if the data is used.}

\item{received_from}{A character string of the date when the benefit
in kind was received written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the earliest date in the data is used.}

\item{received_to}{A character string of the date when the benefit
in kind was received written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the latest date in the data is used.}

\item{registered_from}{A character string of the date when the benefit
in kind was registered written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the earliest date in the data is used.}

\item{registered_to}{A character string of the date when the benefit
in kind was registered written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the latest date in the data is used.}
}
\value{
A tibble of key details on APPG benefits in kind, with one
row per APPG benefit in kind.
}
\description{
\code{appg_benefits} retrieves details on benefits in kind given
to APPGs along with the source of the benefit, benefit value range, date
recieved and registered. Each row is one APPG benefit in kind record.
}
\details{
If no parameters are supplied the function will return a tibble of all
benefits in kind to APPGs.
}
\examples{
appg_benefits()
appg_benefits(category = "Subject", low_value_from = 1000, low_value_to = 2000)
appg_benefits(received_from = "2018-01-01", received_to = "2018-06-01")
}
