% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{appg_financial}
\alias{appg_financial}
\title{Fetch details on APPG financial funding}
\usage{
appg_financial(appg = NA, category = NA, source = NA,
  value_from = NA, value_to = NA, received_from = NA,
  received_to = NA, registered_from = NA, registered_to = NA)
}
\arguments{
\item{appg}{A character string of an APPG title written within a pair of single
qoute or double qoutes. Can be a vector with more than one element. By default
all APPG titles are selected.}

\item{category}{A characer string of a APPG type written within a pair of single
qoute or double qoutes. APPG types available: Country, Subject, Club. Can be a
vector with more than one element. By default all APPG types are selected.}

\item{source}{A character string of the source of financial funding written
within a pair of single qoute or double qoutes. Can be a vector with more than
one element. By default all sources are selected.}

\item{value_from}{A non-negative numeric indicating the value of the financial
funding. By default the minimum value in the data is used.}

\item{value_to}{A non-negative numeric indicating the value of the financial
funding. By default the maximum value if the data is used.}

\item{received_from}{A character string of the date when the financial
funding was received written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the earliest date in the data is used.}

\item{received_to}{A character string of the date when the financial
funding was received written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the latest date in the data is used.}

\item{registered_from}{A character string of the date when the financial
funding was registered written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the earliest date in the data is used.}

\item{registered_to}{A character string of the date when the financial
funding was registered written within a pair of single qoute
or double qoutes. The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD").
By default the latest date in the data is used.}
}
\value{
A tibble of key details on APPG financial funding, with one
row per APPG financial funding.
}
\description{
\code{appg_financial} retrieves details on financial funding given
to APPGs along with the source and value of the funding, date recieved and
registered. Each row is one APPG financial funding record.
}
\details{
If no parameters are supplied the function will return a tibble of all
financial funding given to APPGs.
}
\examples{
appg_financial()
appg_financial(category = "Subject", value_from = 1000, value_to = 2000)
appg_financial(received_from = "2018-01-01", received_to = "2018-06-01")
}
