% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{appg_officers}
\alias{appg_officers}
\title{Fetch details on officers of APPGs}
\usage{
appg_officers(appg = NA, category = NA, officer = NA, party = NA)
}
\arguments{
\item{appg}{A character string of an APPG title written within a pair of single
qoute or double qoutes. Can be a vector with more than one element. By default
all APPG titles are selected.}

\item{category}{A characer string of a APPG type written within a pair of single
qoute or double qoutes. APPG types available: Country, Subject, Club. Can be a
vector with more than one element. By default all APPG types are selected.}

\item{officer}{A character string of an APPG officer written within a pair of
single qoute or double qoutes. Can be a vector with more than one element.
By default all APPG officers are selected.}

\item{party}{A character string of a political party/grouping within the House
of Commons or House of Lords written within a pair of single qoute or double
qoutes. Can be a vector with more than one element. By default all parties are
selected.}
}
\value{
A tibble of details on APPG officers, with one row per APPG
officer.
}
\description{
\code{appg_officers} retrieves details on which MPs and Lords are officers
of APPGs along with their role and party affiliation. Each row is one
APPG officer.
}
\details{
If no parameters are supplied the function will return all APPG officers.
}
\examples{
appg_officers()
appg_officers(officer = c("Chi Onwurah", "Lord Chidgey"))
appg_officers(officer = "Chi Onwurah", category = "Subject")
}
