% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{appgs_agm}
\alias{appgs_agm}
\title{Fetch details on APPG Annual General Meetings}
\usage{
appgs_agm(appg = NA, category = NA, statement = NA,
  latest_agm_from = NA, latest_agm_to = NA, deadline_from = NA,
  deadline_to = NA)
}
\arguments{
\item{appg}{A character string of an APPG title written within a pair of single
qoute or double qoutes. Can be a vector with more than one element. By default
all APPG titles are selected.}

\item{category}{A characer string of a APPG type written within a pair of single
qoute or double qoutes. APPG types available: Country, Subject, Club. Can be a
vector with more than one element. By default all APPG types are selected.}

\item{statement}{A character string of Yes or No as to whether the APPG issued
a financial statment at the latest AGM written within a pair of single qoute
or double qoutes. By default both Yes and No is selected.}

\item{latest_agm_from}{A character string of the date of the latest AGM
written within a pair of single qoute or double qoutes. The date must be
in ISO 8601 format, i.e. ("YYYY-MM-DD"). By default the earliest date in
the data is used.}

\item{latest_agm_to}{A character string of the date of the latest AGM
written within a pair of single qoute or double qoutes. The date must be
in ISO 8601 format, i.e. ("YYYY-MM-DD"). By default the latest date in
the data is used.}

\item{deadline_from}{A character string of the date of the reporting
deadline written within a pair of single qoute or double qoutes.
The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD"). By default the
earliest date in the data is used.}

\item{deadline_to}{A character string of the date of the reporting
deadline written within a pair of single qoute or double qoutes.
The date must be in ISO 8601 format, i.e. ("YYYY-MM-DD"). By default the
latest date in the data is used.}
}
\value{
A tibble of key details on APPG Annual General Meetings, with one
row per APPG AGM.
}
\description{
\code{appgs_agm} retrieves details on APPG Annual General Meetings
along with whether a financial statement was issued, the latest AGM meeting
and reporting year. Each row is one APPG AGM record.
}
\details{
If no parameters are supplied the function will return a tibble of all
APPG AGMs.
}
\examples{
appgs_agm()
appgs_agm(category = "Subject", statement = "No")
appgs_agm(deadline_from = "2019-11-01", deadline_to = "2020-06-01")
}
