% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{appgs_groups}
\alias{appgs_groups}
\title{Fetch details on APPG groups}
\usage{
appgs_groups(appg = NA, category = NA)
}
\arguments{
\item{appg}{A character string of an APPG title written within a pair of single
qoute or double qoutes. Can be a vector with more than one element. By default
all APPG titles are selected.}

\item{category}{A characer string of a APPG type written within a pair of single
qoute or double qoutes. APPG types available: Country, Subject, Club. Can be a
vector with more than one element. By default all APPG types are selected.}
}
\value{
A tibble of basic details of APPGs, with one row per APPG.
}
\description{
\code{appgs_groups} retrieves basic details on the names of APPGs, their
purpose and category type. Each row is one APPG.
}
\details{
If no parameters are supplied the function will return all APPGs.
}
\examples{
appgs_groups()
appgs_groups(category = "Subject")
}
