% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{appgs_members}
\alias{appgs_members}
\title{Fetch details on members of APPGs}
\usage{
appgs_members(appg = NA, category = NA, member = NA, party = NA)
}
\arguments{
\item{appg}{A character string of an APPG title written within a pair of single
qoute or double qoutes. Can be a vector with more than one element. By default
all APPG titles are selected.}

\item{category}{A characer string of a APPG type written within a pair of single
qoute or double qoutes. APPG types available: Country, Subject, Club. Can be a
vector with more than one element. By default all APPG types are selected.}

\item{member}{A character string of an APPG member written within a pair of
single qoute or double qoutes. Can be a vector with more than one element.
By default all APPG members are selected.}

\item{party}{A character string of a political party/grouping within the House
of Commons or House of Lords written within a pair of single qoute or double
qoutes. Can be a vector with more than one element. By default all parties are
selected.}
}
\value{
A tibble of details on APPG membership, with one row per APPG
member.
}
\description{
\code{appgs_members} retrieves details on which MPs and Lords are members
of APPGs along with their role and party affiliation. Each row is one
APPG membership.
}
\details{
If no parameters are supplied the function will return all APPG members.
}
\examples{
appgs_members()
appgs_members(member = c("Chi Onwurah", "Lord Chidgey"))
appgs_members(member = "Chi Onwurah", category = "Subject")
}
