% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{download_appgs}
\alias{download_appgs}
\title{Download Register of All-Party Parliamentary Groups}
\usage{
download_appgs(register_date, pause = 1, save = FALSE)
}
\arguments{
\item{register_date}{A character string of the date of the register of interest
written within a pair of single qoute or double qoutes. The date must be in
ISO 8601 format, i.e. ("YYYY-MM-DD").}

\item{pause}{A non-negative numeric indicating length of time in seconds to
pause between accessing APPG pages. Set to 1 by default. A longer pause helps
minimise burden on the Parliamentary website.}

\item{save}{A boolean which if TRUE saves the cached tables as a series of
.csv files. Additionally, a list object of the raw APPG download is saved as an
.RDS file. All files are saved in the current working environmnent. By default
save is FALSE.}
}
\description{
\code{download_appgs} scrapes and caches entries in the Register of All-Party
Parliamentary Groups. It is not necessary to assign when using download_appgs(),
although if you do a list object of raw tables is returned. If the environment
is closed the cached download will be lost.
}
\examples{
download_appgs("2019-11-05")
download_appgs("2019-11-05", pause = 0.5)
download_appgs("2019-11-05", pause = 0.5, save = TRUE)
data <- download_appgs("2019-11-05")
}
