% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fetch_appg_members}
\alias{fetch_appg_members}
\title{\code{fetch_appg_members} fetches details on which MPs and Lords are members
of APPGs along with their role and party affiliation. Each row is one
APPG membership.}
\usage{
fetch_appg_members(appg = NA, category = NA, member = NA,
  party = NA)
}
\arguments{
\item{appg}{A character string of an APPG title written within a pair of single
qoute or double qoutes. Can be a vector with more than one element. By default
all APPG titles are selected.}

\item{category}{A characer string of a APPG type written within a pair of single
qoute or double qoutes. APPG types available: Country, Subject, Club. Can be a
vector with more than one element. By default all APPG types are selected.}

\item{member}{A character string of an APPG member written within a pair of
single qoute or double qoutes. Can be a vector with more than one element.
By default all APPG members are selected.}

\item{party}{A character string of a political party/grouping within the House
of Commons or House of Lords written within a pair of single qoute or double
qoutes. Can be a vector with more than one element. By default all parties are
selected.}
}
\value{
A tibble of details on APPG membership, with one row per APPG
member.
}
\description{
If no parameters are supplied the function will return all APPG members.
}
\examples{
fetch_appg_members()
fetch_appg_members(member = c("Chi Onwurah", "Lord Chidgey"))
fetch_appg_members(member = c("Chi Onwurah"), category = "Subject")
}
