% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_cached_tables}
\alias{save_cached_tables}
\title{Save cached tables to working directory}
\usage{
save_cached_tables(register_date)
}
\description{
Save cached tables to working directory
}
\keyword{internal}
